Add folders to the .gitignore file
I added the 'dist' and 'node_modules' folders to the .gitignore file.

The 'dist' folder will hold all of the production-ready code that will
be distributed to the server. If a developer wants to recreate the
'dist' folder, they just have to run the code in the GitHub repository.

The 'node_modules' folder will hold all of the installed modules from
node that are used in this project. If a developer wants to recreate the
'node_modules' folder, they just have to run `npm install` where the
package.json file is located.