Add a viewport meta tag to the header
I added a vieport meta tag that lets the browser know to replace the
default viewport width with the width of the device and set the initial
zoom level to one in which the virtual pixels described in the web pages
matches the physical pixels of the device.

This is done to increase the readability and usability of the web site
when viewed on a mobile screen.