Replace the HTML Hint rules for a variable
The HTML Hint rules were being used more than twice in the same
document. They are identical and are not expected to change in the
future. The two set of rules were replaced by a variable that contains
the set of rules, making the file shorter and easier for developers.